requirements.txt
multidict==4.7.6
numpy==1.19.1
openpyxl==3.0.5
python-dateutil==2.8.1
python-Levenshtein==0.12.0
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
